src/pvdss-master/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
contourpy==1.2.1
cx-Oracle==8.3.0
cycler==0.12.1
Cython==0.29.26
decorator==5.1.0
defusedxml==0.7.1
Django==4.2.10
django-auth-ldap==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.0.0
fonttools==4.51.0
future==0.18.3
greenlet==3.0.3
gunicorn==22.0.0
isodate==0.6.1
joblib==1.1.1
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.21.5
packaging==24.0
pandas==1.3.5
pandasql==0.7.3
patsy==0.5.2
pillow==10.3.0
pip==21.3.1
pyagrum==0.22.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==2.0.0
PyJWT==2.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-ldap==3.4.0
pytz==2021.3
scikit-learn==1.0.2
scipy==1.8.0
setuptools==57.4.0
six==1.16.0
SQLAlchemy==1.4.26
sqlparse==0.5.0
statsmodels==0.13.1
threadpoolctl==3.0.0
tqdm==4.62.3
typing_extensions==4.0.1
wheel==0.38.4